+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/BoxTestGenerated.java
@@ -721,6 +721,18 @@ public class BoxTestGenerated extends AbstractBoxTest {
     }
 
     @Test
+    @TestMetadata("IncludesDataClass.kt")
+    public void testIncludesDataClass() {
+      runTest("compiler-tests/src/test/data/box/dependencygraph/IncludesDataClass.kt");
+    }
+
+    @Test
+    @TestMetadata("IncludesDataClassFromExternalModule.kt")
+    public void testIncludesDataClassFromExternalModule() {
+      runTest("compiler-tests/src/test/data/box/dependencygraph/IncludesDataClassFromExternalModule.kt");
+    }
+
+    @Test
     @TestMetadata("IncludesDeepInheritedInterfacesWork.kt")
     public void testIncludesDeepInheritedInterfacesWork() {
       runTest("compiler-tests/src/test/data/box/dependencygraph/IncludesDeepInheritedInterfacesWork.kt");